22: Fix uncaught exception & define __ne__
In np arrays, `foo == bar` yields another np array, which can't be
evaluated as a boolean; it'll raise an exception instead.  So use a np
library call to get the right semantics.

Also `foo != bar` calls `foo.__ne__(bar)` under the hood, which is
defined for all objects in python, so you don't get the call you think
you're getting.