Clear roadblocks to testing of modelforms & models
Automated testing of models is complicated by the tenant system.
TenantTestCase couldn't help because of the extra fields in GeoKitSite,
so make a subclass that covers for that.  Automatic DB setup couldn't
happen because hstore bits were missing from migrations.  All this had
to be done because GeoKitTableForm accesses its model during is_valid();
what I was trying to do was check GeoKitTableForm.clean() by showing it
some csv.

Undocumented gnarly things are harder to test, hence doc comments and
rewrapping of lines.