Fix for ALLOWED_HOSTS schema-detecting bug
New setting for detecting the tenant schema name in the host part of the
request URL, and advice about how to use it in local.  ALLOWED_HOSTS is
really for security, not for application logic, and it has special
documented assumptions we were violating, and it was wrecking automated
testing.

Couple other changes too:  The "*.foo*' entries in ALLOWED_HOSTS don't do
anything so leave them out, also note how to get more output from py.test.